fix: harden FAIR 1.0 foundation
